refactor: move some deperacted compilation methods into deperacted.go
