+++ pkg/api/kpm_pkg_test.go
@@ -6,9 +6,7 @@ import (
 	"testing"
 
 	"gotest.tools/v3/assert"
-	"kcl-lang.io/kcl-go/pkg/kcl"
 	"kcl-lang.io/kpm/pkg/client"
-	"kcl-lang.io/kpm/pkg/opt"
 )
 
 func TestPackageApi(t *testing.T) {
@@ -32,7 +30,7 @@ func TestPackageApi(t *testing.T) {
 	assert.Equal(t, dep.Version, "1.27")
 	assert.Equal(t, dep.Source.Registry.Oci.Reg, "ghcr.io")
 	assert.Equal(t, dep.Source.Registry.Oci.Repo, "kcl-lang/k8s")
-	assert.Equal(t, dep.Source.Registry .Oci.Tag, "1.27")
+	assert.Equal(t, dep.Source.Registry.Oci.Tag, "1.27")
 
 	assert.Equal(t, dep.GetLocalFullPath(""), filepath.Join(kcl_pkg_path, "k8s_1.27"))
 
@@ -185,28 +183,6 @@ func TestGetSchemaTypeUnderEmptyDir(t *testing.T) {
 	assert.Equal(t, len(schemas), 1)
 	assert.Equal(t, schemas[filepath.Join(".")]["SchemaInMain"].Type, "schema")
 	assert.Equal(t, schemas[filepath.Join(".")]["SchemaInMain"].SchemaName, "SchemaInMain")
-}
-
-func TestGetEntries(t *testing.T) {
-	testPath := getTestDir("test_get_entries")
-	pkgPath := filepath.Join(testPath, "no_entries")
-	pkg, err := GetKclPackage(pkgPath)
-	assert.Equal(t, err, nil)
-	assert.Equal(t, len(pkg.GetPkgProfile().GetEntries()), 0)
-
-	pkgPath = filepath.Join(testPath, "with_path_entries")
-	pkg, err = GetKclPackage(pkgPath)
-	assert.Equal(t, err, nil)
-	assert.Equal(t, len(pkg.GetPkgProfile().GetEntries()), 1)
-
-	res, err := RunWithOpts(
-		opt.WithEntries(pkg.GetPkgProfile().GetEntries()),
-		opt.WithKclOption(kcl.WithWorkDir(pkgPath)),
-	)
-
-	assert.Equal(t, err, nil)
-	assert.Equal(t, res.GetRawYamlResult(), "sub: test in sub")
-	assert.Equal(t, res.GetRawJsonResult(), "{\"sub\": \"test in sub\"}")
 }
 
 func TestExportSwaggerV2Spec(t *testing.T) {
