+++ pkg/client/client.go
@@ -490,200 +490,6 @@ func (c *KpmClient) Compile(kclPkg *pkg.KclPkg, kclvmCompiler *runner.Compiler) 
 	return kclvmCompiler.Run()
 }
 
-// CompileWithOpts will compile the kcl program with the compile options.
-func (c *KpmClient) CompileWithOpts(opts *opt.CompileOptions) (*kcl.KCLResultList, error) {
-	pkgPath, err := filepath.Abs(opts.PkgPath())
-	if err != nil {
-		return nil, reporter.NewErrorEvent(reporter.Bug, err, "internal bugs, please contact us to fix it.")
-	}
-
-	c.noSumCheck = opts.NoSumCheck()
-	c.logWriter = opts.LogWriter()
-
-	kclPkg, err := c.LoadPkgFromPath(pkgPath)
-	if err != nil {
-		return nil, err
-	}
-
-	kclPkg.SetVendorMode(opts.IsVendor())
-
-	globalPkgPath, err := env.GetAbsPkgPath()
-	if err != nil {
-		return nil, err
-	}
-
-	err = kclPkg.ValidateKpmHome(globalPkgPath)
-	if err != (*reporter.KpmEvent)(nil) {
-		return nil, err
-	}
-	// add all the options from 'kcl.mod'
-	opts.Merge(*kclPkg.GetKclOpts())
-	if len(opts.Entries()) > 0 {
-		// add entry from '--input'
-		for _, entry := range opts.Entries() {
-			if filepath.IsAbs(entry) {
-				opts.Merge(kcl.WithKFilenames(entry))
-			} else {
-				opts.Merge(kcl.WithKFilenames(filepath.Join(opts.PkgPath(), entry)))
-			}
-		}
-	} else if len(kclPkg.GetEntryKclFilesFromModFile()) == 0 {
-		// No entries profile in kcl.mod and no file settings in the settings file
-		if !opts.HasSettingsYaml() {
-			// No settings file.
-			opts.Merge(kcl.WithKFilenames(opts.PkgPath()))
-		} else if opts.HasSettingsYaml() && len(opts.KFilenameList) == 0 {
-			// Has settings file but no file config in the settings files.
-			opts.Merge(kcl.WithKFilenames(opts.PkgPath()))
-		}
-	}
-	opts.Merge(kcl.WithWorkDir(opts.PkgPath()))
-
-	// Calculate the absolute path of entry file described by '--input'.
-	compiler := runner.NewCompilerWithOpts(opts)
-
-	// Call the kcl compiler.
-	compileResult, err := c.Compile(kclPkg, compiler)
-
-	if err != nil {
-		return nil, reporter.NewErrorEvent(reporter.CompileFailed, err, "failed to compile the kcl package")
-	}
-
-	return compileResult, nil
-}
-
-// RunWithOpts will compile the kcl package with the compile options.
-func (c *KpmClient) RunWithOpts(opts ...opt.Option) (*kcl.KCLResultList, error) {
-	mergedOpts := opt.DefaultCompileOptions()
-	for _, opt := range opts {
-		opt(mergedOpts)
-	}
-	return c.CompileWithOpts(mergedOpts)
-}
-
-// CompilePkgWithOpts will compile the kcl package with the compile options.
-func (c *KpmClient) CompilePkgWithOpts(kclPkg *pkg.KclPkg, opts *opt.CompileOptions) (*kcl.KCLResultList, error) {
-	opts.SetPkgPath(kclPkg.HomePath)
-	if len(opts.Entries()) > 0 {
-		// add entry from '--input'
-		for _, entry := range opts.Entries() {
-			if filepath.IsAbs(entry) {
-				opts.Merge(kcl.WithKFilenames(entry))
-			} else {
-				opts.Merge(kcl.WithKFilenames(filepath.Join(opts.PkgPath(), entry)))
-			}
-		}
-		// add entry from 'kcl.mod'
-	} else if len(kclPkg.GetEntryKclFilesFromModFile()) > 0 {
-		opts.Merge(*kclPkg.GetKclOpts())
-	} else if !opts.HasSettingsYaml() {
-		// no entry
-		opts.Merge(kcl.WithKFilenames(opts.PkgPath()))
-	}
-	opts.Merge(kcl.WithWorkDir(opts.PkgPath()))
-	// Calculate the absolute path of entry file described by '--input'.
-	compiler := runner.NewCompilerWithOpts(opts)
-	// Call the kcl compiler.
-	compileResult, err := c.Compile(kclPkg, compiler)
-
-	if err != nil {
-		return nil, reporter.NewErrorEvent(reporter.CompileFailed, err, "failed to compile the kcl package")
-	}
-
-	return compileResult, nil
-}
-
-// CompileTarPkg will compile the kcl package from the tar package.
-func (c *KpmClient) CompileTarPkg(tarPath string, opts *opt.CompileOptions) (*kcl.KCLResultList, error) {
-	absTarPath, err := utils.AbsTarPath(tarPath)
-	if err != nil {
-		return nil, err
-	}
-	// Extract the tar package to a directory with the same name.
-	// e.g.
-	// 'xxx/xxx/xxx/test.tar' will be extracted to the directory 'xxx/xxx/xxx/test'.
-	destDir := strings.TrimSuffix(absTarPath, filepath.Ext(absTarPath))
-	err = utils.UnTarDir(absTarPath, destDir)
-	if err != nil {
-		return nil, err
-	}
-
-	opts.SetPkgPath(destDir)
-	// The directory after extracting the tar package is taken as the root directory of the package,
-	// and kclvm is called to compile the kcl program under the 'destDir'.
-	// e.g.
-	// if the tar path is 'xxx/xxx/xxx/test.tar',
-	// the 'xxx/xxx/xxx/test' will be taken as the root path of the kcl package to compile.
-	return c.CompileWithOpts(opts)
-}
-
-// CompileGitPkg will compile the kcl package from the git url.
-func (c *KpmClient) CompileGitPkg(gitOpts *git.CloneOptions, compileOpts *opt.CompileOptions) (*kcl.KCLResultList, error) {
-	// 1. Create the temporary directory to pull the tar.
-	tmpDir, err := os.MkdirTemp("", "")
-	if err != nil {
-		return nil, reporter.NewErrorEvent(reporter.Bug, err, "internal bugs, please contact us to fix it.")
-	}
-	tmpDir = filepath.Join(tmpDir, constants.GitEntry)
-
-	// clean the temp dir.
-	defer os.RemoveAll(tmpDir)
-
-	// 2. clone the git repo
-	_, err = git.CloneWithOpts(
-		git.WithCommit(gitOpts.Commit),
-		git.WithBranch(gitOpts.Branch),
-		git.WithTag(gitOpts.Tag),
-		git.WithRepoURL(gitOpts.RepoURL),
-		git.WithLocalPath(tmpDir),
-	)
-	if err != nil {
-		return nil, reporter.NewErrorEvent(reporter.FailedGetPkg, err, "failed to get the git repository")
-	}
-
-	compileOpts.SetPkgPath(tmpDir)
-
-	return c.CompileWithOpts(compileOpts)
-}
-
-// CompileOciPkg will compile the kcl package from the OCI reference or url.
-func (c *KpmClient) CompileOciPkg(ociSource, version string, opts *opt.CompileOptions) (*kcl.KCLResultList, error) {
-	ociOpts, err := c.ParseOciOptionFromString(ociSource, version)
-
-	if err != nil {
-		return nil, err
-	}
-
-	// 1. Create the temporary directory to pull the tar.
-	tmpDir, err := os.MkdirTemp("", "")
-	if err != nil {
-		return nil, reporter.NewErrorEvent(reporter.Bug, err, "internal bugs, please contact us to fix it.")
-	}
-	// clean the temp dir.
-	defer os.RemoveAll(tmpDir)
-
-	localPath := ociOpts.SanitizePathWithSuffix(tmpDir)
-
-	// 2. Pull the tar.
-	err = c.pullTarFromOci(localPath, ociOpts)
-
-	if err != nil {
-		return nil, err
-	}
-
-	// 3.Get the (*.tar) file path.
-	matches, err := filepath.Glob(filepath.Join(localPath, constants.KCL_PKG_TAR))
-	if err != nil || len(matches) != 1 {
-		if err != nil {
-			return nil, reporter.NewErrorEvent(reporter.FailedGetPkg, err, "failed to pull kcl package")
-		} else {
-			return nil, errors.FailedPull
-		}
-	}
-
-	return c.CompileTarPkg(matches[0], opts)
-}
-
 // createIfNotExist will create a file if it does not exist.
 func (c *KpmClient) createIfNotExist(filepath string, storeFunc func() error) error {
 	reporter.ReportMsgTo(fmt.Sprintf("creating new :%s", filepath), c.GetLogWriter())
@@ -1432,86 +1238,6 @@ func (c *KpmClient) DownloadPkgFromOci(dep *downloader.Oci, localPath string) (*
 	}
 
 	return pkg, nil
-}
-
-// DownloadFromOci will download the dependency from the oci repository.
-// Deprecated: Use the DownloadPkgFromOci instead.
-func (c *KpmClient) DownloadFromOci(dep *downloader.Oci, localPath string) (string, error) {
-	ociClient, err := oci.NewOciClient(dep.Reg, dep.Repo, &c.settings)
-	if err != nil {
-		return "", err
-	}
-	ociClient.SetLogWriter(c.logWriter)
-	// Select the latest tag, if the tag, the user inputed, is empty.
-	var tagSelected string
-	if len(dep.Tag) == 0 {
-		tagSelected, err = ociClient.TheLatestTag()
-		if err != nil {
-			return "", err
-		}
-
-		reporter.ReportMsgTo(
-			fmt.Sprintf("the lastest version '%s' will be added", tagSelected),
-			c.logWriter,
-		)
-
-		dep.Tag = tagSelected
-		localPath = localPath + dep.Tag
-	} else {
-		tagSelected = dep.Tag
-	}
-
-	reporter.ReportMsgTo(
-		fmt.Sprintf("downloading '%s:%s' from '%s/%s:%s'", dep.Repo, tagSelected, dep.Reg, dep.Repo, tagSelected),
-		c.logWriter,
-	)
-
-	// Pull the package with the tag.
-	err = ociClient.Pull(localPath, tagSelected)
-	if err != nil {
-		return "", err
-	}
-
-	matches, _ := filepath.Glob(filepath.Join(localPath, "*.tar"))
-	if matches == nil || len(matches) != 1 {
-		// then try to glob tgz file
-		matches, _ = filepath.Glob(filepath.Join(localPath, "*.tgz"))
-		if matches == nil || len(matches) != 1 {
-			return "", reporter.NewErrorEvent(
-				reporter.InvalidKclPkg,
-				err,
-				fmt.Sprintf("failed to find the kcl package from '%s'.", localPath),
-			)
-		}
-	}
-
-	tarPath := matches[0]
-	if utils.IsTar(tarPath) {
-		err = utils.UnTarDir(tarPath, localPath)
-	} else {
-		err = utils.ExtractTarball(tarPath, localPath)
-	}
-	if err != nil {
-		return "", reporter.NewErrorEvent(
-			reporter.FailedUntarKclPkg,
-			err,
-			fmt.Sprintf("failed to untar the kcl package from '%s' into '%s'.", tarPath, localPath),
-		)
-	}
-
-	// After untar the downloaded kcl package tar file, remove the tar file.
-	if utils.DirExists(tarPath) {
-		rmErr := os.Remove(tarPath)
-		if rmErr != nil {
-			return "", reporter.NewErrorEvent(
-				reporter.FailedUntarKclPkg,
-				err,
-				fmt.Sprintf("failed to untar the kcl package tar from '%s' into '%s'.", tarPath, localPath),
-			)
-		}
-	}
-
-	return localPath, nil
 }
 
 // PullFromOci will pull a kcl package from oci registry and unpack it.
