+++ pkg/client/test_data/test_run_git/expect.yaml
@@ -1,0 +1,39 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: flask-demo
+  labels:
+    app: flask-demo
+  namespace: default
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: flask-demo
+  template:
+    metadata:
+      labels:
+        app: flask-demo
+    spec:
+      containers:
+      - name: flaskdemo
+        image: kcllang/flask_demo:8d31498e765ff67a2fa9933d4adffe067544b2fe
+        ports:
+        - protocol: TCP
+          containerPort: 5000
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: flask-demo
+  labels:
+    app: flask-demo
+  namespace: default
+spec:
+  type: NodePort
+  selector:
+    app: flask-demo
+  ports:
+  - port: 5000
+    protocol: TCP
+    targetPort: 5000
