--- pkg/client/test_data/test_run_git/expect1.json
@@ -1,11 +1,0 @@
-{
-    "hellocollaset": {
-        "workload": {
-            "containers": {
-                "nginx": {
-                    "image": "nginx:v2"
-                }
-            }
-        }
-    }
-}
