--- pkg/client/test_data/test_run_git/expect3.json
@@ -1,35 +1,0 @@
-{
-    "wordpress": {
-        "workload": {
-            "containers": {
-                "wordpress": {
-                    "image": "wordpress:6.3",
-                    "env": {
-                        "WORDPRESS_DB_HOST": "$(KUSION_DB_HOST_WORDPRESS)",
-                        "WORDPRESS_DB_USER": "$(KUSION_DB_USERNAME_WORDPRESS)",
-                        "WORDPRESS_DB_PASSWORD": "$(KUSION_DB_PASSWORD_WORDPRESS)",
-                        "WORDPRESS_DB_NAME": "mysql"
-                    },
-                    "resources": {
-                        "cpu": "500m",
-                        "memory": "512Mi"
-                    }
-                }
-            },
-            "replicas": 1,
-            "ports": [
-                {
-                    "port": 80,
-                    "protocol": "TCP",
-                    "public": false
-                }
-            ]
-        },
-        "database": {
-            "wordpress": {
-                "type": "cloud",
-                "version": "8.0"
-            }
-        }
-    }
-}
