+++ test/e2e/test_suites/kpm/exec_inside_pkg/run_with_not_exist_input/test_suite.stderr
@@ -1,2 +1,1 @@
-failed to compile the kcl package
-Cannot find the kcl file, please check the file path
+Cannot find the kcl file, please check the file path <workspace>/test_kcl/no_exist
