--- test/e2e/test_suites/kpm/exec_outside_pkg/run_tar_with_noargs_input/test_suite.env
@@ -1,2 +1,0 @@
-KPM_HOME=""
-KCLVM_VENDOR_HOME=""
