--- test/e2e/test_suites/kpm/exec_outside_pkg/run_tar_with_noargs_input/test_suite.input
@@ -1,1 +1,0 @@
-kpm run --input kcl1.tar
