--- test/e2e/test_suites/kpm/exec_outside_pkg/run_tar_with_not_exist_input/test_suite.input
@@ -1,1 +1,0 @@
-kpm run --input no_exist --vendor kcl1.tar
