--- test/e2e/test_suites/kpm/exec_outside_pkg/run_tar_with_not_exist_input/test_suite.stderr
@@ -1,2 +1,0 @@
-failed to compile the kcl package
-Cannot find the kcl file, please check the file path
