+++ test/e2e/test_suites/kpm/exec_outside_pkg/run_tar_with_not_exist_tar/test_suite.stderr
@@ -1,1 +1,2 @@
-the kcl package tar path is not found
+failed to open 'no_exist.tar'
+open no_exist.tar: no such file or directory
