--- test/e2e/test_suites/kpm/exec_outside_pkg/run_with_no_args/test_suite.input
@@ -1,1 +1,0 @@
-kpm run
