--- test/e2e/test_suites/kpm/exec_outside_pkg/run_with_no_args/test_suite.stderr
@@ -1,2 +1,0 @@
-could not load 'kcl.mod' in '<workspace>'
-open <workspace>/kcl.mod: no such file or directory
