+++ test/e2e/test_suites/kpm/kpm_run/test_kpm_run_duplication_deps/test_suite.stderr
@@ -1,4 +1,3 @@
-failed to compile the kcl package
 because '-' in the original dependency names is replaced with '_'
 please check your dependencies with '-' or '_' in dependency name
 dependency name conflict, 'dep_with_line' already exists
